profileHive

+++ lib/features/Profile/data/repo_impl/repo_impl.dart
@@ -8,17 +8,30 @@ import 'package:meal_recommendation_b1/core/networking/firebase_error_model.dart
 import '../../domain/entity/entity.dart';
 import '../../domain/repo/repo.dart';
 import '../Model/UserModel.dart';
+import '../dataSource/local/LocalData.dart';
 
 class FirebaseUserRepository implements UserRepository {
   final FirebaseFirestore firestore;
+  final HiveLocalUserDataSource localDataSource;
 
-  FirebaseUserRepository(this.firestore);
+  FirebaseUserRepository(this.firestore, this.localDataSource);
 
   @override
   Future<User> getUserProfile(String userId) async {
+    // Check local storage first
+    final localUser = localDataSource.getUser(userId);
+    if (localUser != null) {
+      return localUser;
+    }
+
+    // If not in local storage, fetch from Firebase
     final doc = await firestore.collection('users').doc(userId).get();
     if (doc.exists) {
-      return UserModel.fromFirestore(doc.data()!, doc.id);
+      final user = UserModel.fromFirestore(doc.data()!, doc.id);
+
+      // Save to local storage for future access
+      await localDataSource.saveUser(user);
+      return user;
     } else {
       throw Exception("User not found");
     }
@@ -30,56 +43,62 @@ class FirebaseUserRepository implements UserRepository {
         .collection('users')
         .doc(user.id)
         .update((user).toFirestore());
-    await getUserProfile(user.id);
+
+    // Update the local storage
+    await localDataSource.saveUser(user);
   }
 
   @override
-  Future<Either<FirebaseErrorModel, String>> uploadUserImage(
-      {required File newImage, String? oldImage}) async {
+  Future<Either<FirebaseErrorModel, String>> uploadUserImage({
+    required File newImage,
+    String? oldImage,
+  }) async {
     return await _uploadImageToFirebaseStorage(
-        image: newImage, oldImage: oldImage);
+      image: newImage,
+      oldImage: oldImage,
+    );
   }
 
-  Future<Either<FirebaseErrorModel, String>> _uploadImageToFirebaseStorage(
-      {required File image, String? oldImage}) async {
+  Future<Either<FirebaseErrorModel, String>> _uploadImageToFirebaseStorage({
+    required File image,
+    String? oldImage,
+  }) async {
     try {
       String filePath = 'profile_images/${DateTime.now()}.png';
-      final imageUrl =
-          await FirebaseStorage.instance.ref(filePath).putFile(image);
-      String downlownUrl = await imageUrl.ref.getDownloadURL();
-      if (oldImage!.isNotEmpty) {
+      final imageUrl = await FirebaseStorage.instance.ref(filePath).putFile(image);
+      String downloadUrl = await imageUrl.ref.getDownloadURL();
+
+      if (oldImage != null && oldImage.isNotEmpty) {
         final ref = FirebaseStorage.instance.refFromURL(oldImage);
         await ref.delete();
       }
-      return Right(downlownUrl);
+      return Right(downloadUrl);
     } on FirebaseException catch (e) {
-      // Firebase-specific errors
-      return Left(
-        FirebaseErrorHandler.handle(e),
-      );
+      return Left(FirebaseErrorHandler.handle(e));
     } on SocketException {
-      // Network error
       return Left(
         FirebaseErrorModel(
-            message: 'Network error: No Internet connection.', errorCode: -1),
+            message: 'Network error: No Internet connection.',
+            errorCode: -1
+        ),
       );
     } on TimeoutException {
-      // Handle timeout
       return Left(
         FirebaseErrorModel(message: 'Upload timed out.', errorCode: -1),
       );
     } on FormatException {
-      // Handle format errors
       return Left(
         FirebaseErrorModel(
-            message: 'File format error: Image file format is not supported.',
-            errorCode: -1),
+          message: 'File format error: Image file format is not supported.',
+          errorCode: -1,
+        ),
       );
     } catch (e) {
-      // Generic error handling
       return Left(
         FirebaseErrorModel(
-            message: 'An unexpected error occurred: $e', errorCode: -1),
+          message: 'An unexpected error occurred: $e',
+          errorCode: -1,
+        ),
       );
     }
   }
